requirements.txt
click==7.1.2
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
joblib==1.0.0
lxml==4.6.2
MarkupSafe==1.1.1
murmurhash==1.0.5
nltk==3.5
numpy==1.19.4
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.5
pandas==1.1.5
Pillow==8.0.1
pkg-resources==0.0.0
preshed==3.0.5
pyahocorasick==1.4.0
pyate==0.4.2
pycryptodome==3.9.9
PyPDF2==1.26.0
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.4
regex==2020.11.13
requests==2.25.1
RESTfulEf==0.1.1
scikit-learn==0.24.0
scipy==1.5.4
sortedcontainers==2.3.0
soupsieve==2.1
spacy==2.3.5
SpeechRecognition==3.8.1
textract==1.6.3
thinc==7.4.5
threadpoolctl==2.1.0
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
XlsxWriter==1.3.7
zipp==3.4.0